+++ include/fastrtps/attributes/all_attributes.h
@@ -23,8 +23,8 @@
 #include "PublisherAttributes.h"
 #include "SubscriberAttributes.h"
 #include "TopicAttributes.h"
-#include "RequesterAttributes.h"
-#include "ReplierAttributes.h"
+#include "RequesterAttributes.hpp"
+#include "ReplierAttributes.hpp"
 
 
 #endif /* ALL_ATTRIBUTES_H_ */
