+++ src/cpp/xmlparser/XMLParser.cpp
@@ -2377,7 +2377,7 @@ XMLP_ret XMLParser::parseXMLReplierProf(tinyxml2::XMLElement* p_root, BaseNode& 
     }
     else
     {
-        logError(XMLPARSER, "Error parsing requester profile");
+        logError(XMLPARSER, "Error parsing replier profile");
         ret = XMLP_ret::XML_ERROR;
     }
     return ret;
@@ -2926,7 +2926,7 @@ XMLP_ret XMLParser::fillDataNode(tinyxml2::XMLElement* p_profile, DataNode<Subsc
 XMLP_ret XMLParser::fillDataNode(tinyxml2::XMLElement* p_profile, DataNode<RequesterAttributes>& requester_node)
 {
     /*
-        <xs:complexType name="requesterAttributesType">
+        <xs:complexType name="requesterProfileType">
             <xs:all>
                 <xs:element name="request_topic_name" type="stringType" minOccurs="0"/>
                 <xs:element name="reply_topic_name" type="stringType" minOccurs="0"/>
@@ -3039,7 +3039,7 @@ XMLP_ret XMLParser::fillDataNode(tinyxml2::XMLElement* p_profile, DataNode<Reque
 XMLP_ret XMLParser::fillDataNode(tinyxml2::XMLElement* p_profile, DataNode<ReplierAttributes>& replier_node)
 {
     /*
-        <xs:complexType name="replierAttributesType">
+        <xs:complexType name="replierProfileType">
             <xs:all>
                 <xs:element name="request_topic_name" type="stringType" minOccurs="0"/>
                 <xs:element name="reply_topic_name" type="stringType" minOccurs="0"/>
